add world and its test, refactor of other tests

+++ Prati/Components/Vector2D.cs
@@ -5,7 +5,7 @@ namespace Components
         public double X { get; set;}
         public double Y { get; set;}
 
-        public Vector2D(int x, int y)
+        public Vector2D(double x, double y)
         {
             X = x;
             Y = y;
